Develop pre alpha

+++ gist/Newton_10_fitDirectory.sh
@@ -1,20 +1,37 @@
 # _Newton_10_fitDirectory
 ## fitディレクトリにまとめ
-echo ${My_Newton_D:=$(pwd)}
+FLAG_hardCopy=false # arg
+FLAG_symbLink=false # arg
+tmp_prefix="newton_" # arg
+if [[ $(declare --help | grep -c -o -E "\-g\s+create global variables") -eq 0 ]]; then 
+    My_Newton_D=${My_Newton_D:=$(pwd)} 
+else 
+    declare -g My_Newton_D=${My_Newton_D:=$(pwd)} 
+fi
 cd $My_Newton_D
-tmp_prefix=newton_
 mkdir -p $My_Newton_D/fit $My_Newton_D/../fit/
-
 obs_dirs=($(find . -maxdepth 1 -type d -printf "%P\n" | grep ^[0-9]))
 for My_Newton_ID in ${obs_dirs[@]}; do
-    cp $My_Newton_D/$My_Newton_ID/ODF/fit/${tmp_prefix}* ${My_Newton_D}/fit/ -f
+    if [[ ${FLAG_symbLink:=false} == "true" ]]; then
+        find $My_Newton_D/$My_Newton_ID/ODF/fit/ -name "${tmp_prefix}*.*" \
+            -type f -printf "%f\n" |
+            xargs -n 1 -i rm -f $My_Newton_D/fit/{}
+        ln -s $My_Newton_D/$My_Newton_ID/ODF/fit/${tmp_prefix}* ${My_Newton_D}/fit/
+    else
+        cp -f $My_Newton_D/$My_Newton_ID/ODF/fit/${tmp_prefix}* ${My_Newton_D}/fit/
+    fi
 done
-### remove the files with the same name as new files
-find $My_Newton_D/fit/ -name "${tmp_prefix}*.*" \
-    -type f -printf "%f\n" |
-    xargs -n 1 -i rm -f $My_Newton_D/../fit/{}
-### remove broken symbolic links
+if [[ ${FLAG_hardCopy:=false} == "true" ]]; then
+    cp -f $My_Newton_D/fit/${tmp_prefix}*.* $My_Newton_D/../fit/
+else
+        # remove the files with the same name as new files
+    find $My_Newton_D/fit/ -name "${tmp_prefix}*.*" \
+        -type f -printf "%f\n" |
+        xargs -n 1 -i rm -f $My_Newton_D/../fit/{}
+    # generate symbolic links
+    ln -s $My_Newton_D/fit/${tmp_prefix}*.* $My_Newton_D/../fit/
+fi
+# remove broken symbolic links
 find -L $My_Newton_D/../fit/ -type l -delete
-### generate symbolic links
-ln -s $My_Newton_D/fit/${tmp_prefix}*.* $My_Newton_D/../fit/
+
 cd $My_Newton_D
