+++ gist/SuzakuXis_7_fitDirectory.sh
@@ -1,20 +1,37 @@
 # _SuzakuXis_7_fitDirectory
 ## fitディレクトリにまとめ
-echo ${My_Suzaku_D:=$(pwd)}
+FLAG_hardCopy=false # arg
+FLAG_symbLink=false # arg
+tmp_prefix="xis_" # arg
+if [[ $(declare --help | grep -c -o -E "\-g\s+create global variables") -eq 0 ]]; then 
+    My_Suzaku_D=${My_Suzaku_D:=$(pwd)} 
+else 
+    declare -g My_Suzaku_D=${My_Suzaku_D:=$(pwd)} 
+fi
 cd $My_Suzaku_D
-tmp_prefix=xis_
-obs_dirs=($(find . -maxdepth 1 -type d -printf "%P\n" | grep ^[0-9]))
 mkdir -p $My_Suzaku_D/fit $My_Suzaku_D/../fit
+obs_dirs=($(find . -maxdepth 1 -type d -printf "%P\n" | grep ^[0-9]))
 for My_Suzaku_ID in ${obs_dirs[@]}; do
-    cp $My_Suzaku_D/$My_Suzaku_ID/xis/event_cl/fit/${tmp_prefix}*.* $My_Suzaku_D/fit/ -f
+    if [[ ${FLAG_symbLink:=false} == "true" ]]; then
+        find $My_Suzaku_D/$My_Suzaku_ID/xis/event_cl/fit/ -name "${tmp_prefix}*.*" \
+            -type f -printf "%f\n" |
+            xargs -n 1 -i rm -f $My_Suzaku_D/fit/{}
+        ln -s $My_Suzaku_D/$My_Suzaku_ID/xis/event_cl/fit/${tmp_prefix}* ${My_Suzaku_D}/fit/
+    else
+        cp -f $My_Suzaku_D/$My_Suzaku_ID/xis/event_cl/fit/${tmp_prefix}* ${My_Suzaku_D}/fit/
+    fi
 done
-### remove the files with the same name as new files
-find $My_Suzaku_D/fit/ -name "${tmp_prefix}*.*" \
-    -type f -printf "%f\n" |
-    xargs -n 1 -i rm -f $My_Suzaku_D/../fit/{}
-### remove broken symbolic links
+if [[ ${FLAG_hardCopy:=false} == "true" ]]; then
+    cp -f $My_Suzaku_D/fit/${tmp_prefix}*.* $My_Suzaku_D/../fit/
+else
+        # remove the files with the same name as new files
+    find $My_Suzaku_D/fit/ -name "${tmp_prefix}*.*" \
+        -type f -printf "%f\n" |
+        xargs -n 1 -i rm -f $My_Suzaku_D/../fit/{}
+    # generate symbolic links
+    ln -s $My_Suzaku_D/fit/${tmp_prefix}*.* $My_Suzaku_D/../fit/
+fi
+# remove broken symbolic links
 find -L $My_Suzaku_D/../fit/ -type l -delete
-### generate symbolic links
-ln -s $My_Suzaku_D/fit/${tmp_prefix}*.* $My_Suzaku_D/../fit/
 
 cd $My_Suzaku_D
